WIP: Add SRID support to JSON interchange schema
This sketches out how we could add SRID support to the interchange
schema, allowing WGS 84 (EPSG:4326) to be handled differently from other
SRIDs, which in turn means we would no longer need to covert to a
standard SRID when exporting.

The only missing bit here is in the PG driver, which would need to dig
around in PostGIS-specific tables to get the SRID for `geometry`
columns.